#4 Change Button Component name
